The Threat transparency changed
Alpha up to 0.03 from 0.02

Thought-process: We playtested and people said that the Threat was too hard to see, which we thought was funny - then we actually considered the idea that our playtesters might be right. The decision was made to up the Alpha by exactly .01 to make it easier to see the Threat up close, and a good side effect of this is that you can now see their red eyes up close, adding to the horror effect.